update index/breaking news eScience entries (#49)
* News history (#44)

* first draft update for news archive

* single line tweak to test indenting 

remove a single <ul> to see effects on indentation (or not) following the nested Fellowship items

* two changes for line spacing and indentations

only remove two different ul operations

* two line changes for line spacing & formatting

remove two ul operations, to remove extra line of space, while maintaining left margin alignment

* change main page header

change from 'Old News' to 'ReproNim News Archive' - purely a wording change, but in the process, see whether centering changes or not

* change section header

change header from 'News Archive' to Past Events and Activities- purely a wording change, but will also be a test of whether spacing is maintained

* remove the UPDATE line, text only

strictly a wording deletion, of only the text on this line (71) without removing the start and end paragraph operations, to see whether space allocation is preserved

* remove paragraph at line 71

just remove start and stop of paragraph, which is currently empty, to see if it affects allocation of blank vertical white space on page

* remove text only for update page subheader

remove words only, leave rest of line intact- to remove text not needed, and see whether spacing/formatting is affected in the process

* Remove 2 update statements

remove text only for update statements that = extraneous wording at bottom of viewable page- but leave rest of each line (124, 126) intact

* removed empty header

* Breaking news updates (#45)

* News history (#44)

* first draft update for news archive

* single line tweak to test indenting 

remove a single <ul> to see effects on indentation (or not) following the nested Fellowship items

* two changes for line spacing and indentations

only remove two different ul operations

* two line changes for line spacing & formatting

remove two ul operations, to remove extra line of space, while maintaining left margin alignment

* change main page header

change from 'Old News' to 'ReproNim News Archive' - purely a wording change, but in the process, see whether centering changes or not

* change section header

change header from 'News Archive' to Past Events and Activities- purely a wording change, but will also be a test of whether spacing is maintained

* remove the UPDATE line, text only

strictly a wording deletion, of only the text on this line (71) without removing the start and end paragraph operations, to see whether space allocation is preserved

* remove paragraph at line 71

just remove start and stop of paragraph, which is currently empty, to see if it affects allocation of blank vertical white space on page

* remove text only for update page subheader

remove words only, leave rest of line intact- to remove text not needed, and see whether spacing/formatting is affected in the process

* Remove 2 update statements

remove text only for update statements that = extraneous wording at bottom of viewable page- but leave rest of each line (124, 126) intact

* removed empty header

* Breaking News - Add ReproNim at OHBM 

intended to link to ReproNim Events page, and to OHBM web site

* Adjust line to modify indentation

remove ul from line 106, goal of restoring left margin news items alignment

* fix links

* add news history link

* Adjust text link to News History

* shorten News History link line

add bold

* add blank lines before News History

but w/out li indicator, so may or not work

* remove period

* Added 'tools' section

Added as 'section-6. Currently a pointer to the google doc, which will be webified at some point. Worth debating if this is the best place to put this content...

* fixed mixed content errors

* restored Yarik and Kyle's images

* better image for Al

* removed events code from 5 steps pages

* Remove old Breaking News items, add new ones

* Edits on Breaking News -reserach objects entry -to correct formatting

* formatting change Breaking News entry

* more formatting breaking news

* additional formatting breaking news

* formatting breaking news

* add shareable version google doc for INCF events

* added link to list signup form

* update breaking news eScience events

* news - formatting for martone keynote entry

* news- tweak to back out of bullet point tier

* news formatting revision bullet points

* tweak to remove last unused bullet