Julie test front page links team and collaborators (#257)
* test add links for About Us and Collaboratos

* Additional placement of links to Team and Collabs

testing different placements for linkages, on front page. This is currently an additional option, in line 71. 
The placement in line 89 is still present for now.

* test-Additional option for Team and Collab links

adding 3rd option for team and collaborator links, line 70, just below 'welcome to ReproNim'
using <h3> header to try and create a centered line.
other options still remain for now

* remove h3 header line 70

disrupts other spacing in section, and is out of scale, font size-wise.

* test add links to lines 89-91 alternate option

Different attempt to add Team and Collaborator links to face page
add within a paragraph but separate line to try and offset content.
probably improperly formatted

* remove 'Meet the team...from line 89

Redundant, so leaving it in line 91 for now and removing from 89

* remove pointer to Team and Collaborators line 71